requirements.txt
torch>=2.6.0
diffusers>=0.25.0
transformers>=4.48.0
accelerate>=0.24.0
gradio>=5.23.0
pillow>=10.3.0
safetensors>=0.4.0
